+++ buildbase.py
@@ -433,19 +433,19 @@ def install_webrtc(version, source_dir, install_dir, platform: str):
     extract(archive, output_dir=install_dir, output_dirname="webrtc")
 
 
-def build_webrtc(platform, webrtc_build_dir, webrtc_build_args, debug):
-    with cd(webrtc_build_dir):
+def build_webrtc(platform, local_webrtc_build_dir, local_webrtc_build_args, debug):
+    with cd(local_webrtc_build_dir):
         args = ["--webrtc-nobuild-ios-framework", "--webrtc-nobuild-android-aar"]
         if debug:
             args += ["--debug"]
 
-        args += webrtc_build_args
+        args += local_webrtc_build_args
 
         cmd(["python3", "run.py", "build", platform, *args])
 
         # インクルードディレクトリを増やしたくないので、
         # __config_site を libc++ のディレクトリにコピーしておく
-        webrtc_source_dir = os.path.join(webrtc_build_dir, "_source", platform, "webrtc")
+        webrtc_source_dir = os.path.join(local_webrtc_build_dir, "_source", platform, "webrtc")
         src_config = os.path.join(
             webrtc_source_dir, "src", "buildtools", "third_party", "libc++", "__config_site"
         )
@@ -486,11 +486,11 @@ class WebrtcInfo(NamedTuple):
 
 
 def get_webrtc_info(
-    platform: str, webrtc_build_dir: Optional[str], install_dir: str, debug: bool
+    platform: str, local_webrtc_build_dir: Optional[str], install_dir: str, debug: bool
 ) -> WebrtcInfo:
     webrtc_install_dir = os.path.join(install_dir, "webrtc")
 
-    if webrtc_build_dir is None:
+    if local_webrtc_build_dir is None:
         return WebrtcInfo(
             version_file=os.path.join(webrtc_install_dir, "VERSIONS"),
             deps_file=os.path.join(webrtc_install_dir, "DEPS"),
@@ -501,15 +501,17 @@ def get_webrtc_info(
             libcxx_dir=os.path.join(install_dir, "llvm", "libcxx"),
         )
     else:
-        webrtc_build_source_dir = os.path.join(webrtc_build_dir, "_source", platform, "webrtc")
+        webrtc_build_source_dir = os.path.join(
+            local_webrtc_build_dir, "_source", platform, "webrtc"
+        )
         configuration = "debug" if debug else "release"
         webrtc_build_build_dir = os.path.join(
-            webrtc_build_dir, "_build", platform, configuration, "webrtc"
+            local_webrtc_build_dir, "_build", platform, configuration, "webrtc"
         )
 
         return WebrtcInfo(
-            version_file=os.path.join(webrtc_build_dir, "VERSION"),
-            deps_file=os.path.join(webrtc_build_dir, "DEPS"),
+            version_file=os.path.join(local_webrtc_build_dir, "VERSION"),
+            deps_file=os.path.join(local_webrtc_build_dir, "DEPS"),
             webrtc_include_dir=os.path.join(webrtc_build_source_dir, "src"),
             webrtc_source_dir=os.path.join(webrtc_build_source_dir, "src"),
             webrtc_library_dir=webrtc_build_build_dir,
@@ -533,6 +535,101 @@ def install_boost(version, source_dir, install_dir, sora_version, platform: str)
     )
     rm_rf(os.path.join(install_dir, "boost"))
     extract(archive, output_dir=install_dir, output_dirname="boost")
+
+
+# 以下の問題を解決するためのパッチ
+#
+# No support for msvc-toolset 14.4x (VS 2022, 17.10.x): https://github.com/boostorg/boost/issues/914
+BOOST_PATCH_SUPPORT_14_4 = r"""
+diff --git a/tools/build/src/engine/config_toolset.bat b/tools/build/src/engine/config_toolset.bat
+index 4ba577cac..3e3f6a3a1 100644
+--- a/tools/build/src/engine/config_toolset.bat
++++ b/tools/build/src/engine/config_toolset.bat
+@@ -157,7 +157,7 @@ pushd %CD%
+ if "_%VSINSTALLDIR%_" == "__" call :Call_If_Exists "%B2_TOOLSET_ROOT%Auxiliary\Build\vcvarsall.bat" %B2_BUILD_ARGS%
+ popd
+ @REM set "B2_CXX="%CXX%" /nologo /MP /MT /TP /Feb2 /wd4996 /O2 /GL /EHsc"
+-set "B2_CXX="%CXX%" /nologo -TP /wd4996 /wd4675 /EHs /GR /Zc:throwingNew /O2 /Ob2 /W3 /MD /Zc:forScope /Zc:wchar_t /Zc:inline /Gw /favor:blend /Feb2"
++set "B2_CXX="%CXX%" /nologo -TP /wd4996 /wd4675 /EHs /GR /Zc:throwingNew /O2 /Ob2 /W3 /MT /Zc:forScope /Zc:wchar_t /Zc:inline /Gw /favor:blend /Feb2"
+ set "B2_CXX_LINK=/link kernel32.lib advapi32.lib user32.lib"
+ set "_known_=1"
+ goto :Embed_Minafest_Via_Link
+diff --git a/tools/build/src/tools/msvc.jam b/tools/build/src/tools/msvc.jam
+index 54a6ced32..4bb3810b3 100644
+--- a/tools/build/src/tools/msvc.jam
++++ b/tools/build/src/tools/msvc.jam
+@@ -1137,7 +1137,15 @@ local rule generate-setup-cmd ( version : command : parent : options * : cpu : g
+         }
+         else
+         {
+-            if [ MATCH "(14.3)" : $(version) ]
++            if [ MATCH "(14.4)" : $(version) ]
++            {
++                if $(.debug-configuration)
++                {
++                    ECHO "notice: [generate-setup-cmd] $(version) is 14.4" ;
++                }
++                parent = [ path.native [ path.join  $(parent) "..\\..\\..\\..\\..\\Auxiliary\\Build" ] ] ;
++            }
++            else if [ MATCH "(14.3)" : $(version) ]
+             {
+                 if $(.debug-configuration)
+                 {
+@@ -1316,7 +1324,11 @@ local rule configure-really ( version ? : options * )
+             # version from the path.
+             # FIXME: We currently detect both Microsoft Visual Studio 9.0 and
+             # 9.0express as 9.0 here.
+-            if [ MATCH "(MSVC\\\\14.3)" : $(command) ]
++            if [ MATCH "(MSVC\\\\14.4)" : $(command) ]
++            {
++                version = 14.4 ;
++            }
++            else if [ MATCH "(MSVC\\\\14.3)" : $(command) ]
+             {
+                 version = 14.3 ;
+             }
+@@ -1745,13 +1757,17 @@ local rule default-path ( version )
+         # And fortunately, forward slashes do also work in native Windows.
+         local vswhere = "$(root)/Microsoft Visual Studio/Installer/vswhere.exe" ;
+         # The check for $(root) is to avoid a segmentation fault if not found.
+-        if $(version) in 14.1 14.2 14.3 default && $(root) && [ path.exists $(vswhere) ]
++        if $(version) in 14.1 14.2 14.3 14.4 default && $(root) && [ path.exists $(vswhere) ]
+         {
+             local req = "-requires Microsoft.VisualStudio.Component.VC.Tools.x86.x64" ;
+             local prop = "-property installationPath" ;
+             local limit ;
+ 
+-            if $(version) = 14.3
++            if $(version) = 14.4
++            {
++                limit = "-version \"[17.0,18.0)\" -prerelease" ;
++            }
++            else if $(version) = 14.3
+             {
+                 limit = "-version \"[17.0,18.0)\" -prerelease" ;
+             }
+@@ -2174,7 +2190,7 @@ for local arch in [ MATCH "^\\.cpus-on-(.*)" : [ VARNAMES $(__name__) ] ]
+                      armv7 armv7s ;
+ 
+ # Known toolset versions, in order of preference.
+-.known-versions = 14.3 14.2 14.1 14.0 12.0 11.0 10.0 10.0express 9.0 9.0express 8.0 8.0express 7.1
++.known-versions = 14.4 14.3 14.2 14.1 14.0 12.0 11.0 10.0 10.0express 9.0 9.0express 8.0 8.0express 7.1
+     7.1toolkit 7.0 6.0 ;
+ 
+ # Version aliases.
+@@ -2226,6 +2242,11 @@ for local arch in [ MATCH "^\\.cpus-on-(.*)" : [ VARNAMES $(__name__) ] ]
+     "Microsoft Visual Studio/2022/*/VC/Tools/MSVC/*/bin/Host*/*"
+     ;
+ .version-14.3-env = VS170COMNTOOLS ProgramFiles ProgramFiles(x86) ;
++.version-14.4-path =
++    "../../VC/Tools/MSVC/*/bin/Host*/*"
++    "Microsoft Visual Studio/2022/*/VC/Tools/MSVC/*/bin/Host*/*"
++    ;
++.version-14.4-env = VS170COMNTOOLS ProgramFiles ProgramFiles(x86) ;
+ 
+ # Auto-detect all the available msvc installations on the system.
+ auto-detect-toolset-versions ;
+"""
 
 
 @versioned
@@ -563,6 +660,25 @@ def build_and_install_boost(
         bootstrap = ".\\bootstrap.bat" if target_os == "windows" else "./bootstrap.sh"
         b2 = "b2" if target_os == "windows" else "./b2"
         runtime_link = "static" if target_os == "windows" else "shared"
+
+        # Windows かつ Boost 1.85.0 の場合はパッチを当てる
+        if target_os == "windows" and version == "1.85.0":
+            directory = cmdcap(["git", "rev-parse", "--show-prefix"])
+            cmd(
+                [
+                    "git",
+                    "apply",
+                    "-p1",
+                    "--ignore-space-change",
+                    "--ignore-whitespace",
+                    "--whitespace=nowarn",
+                    f"--directory={directory}",
+                    "-",
+                ],
+                input=BOOST_PATCH_SUPPORT_14_4,
+                text=True,
+                encoding="utf-8",
+            )
 
         cmd([bootstrap])
 
@@ -749,15 +865,23 @@ def install_sora_and_deps(platform: str, source_dir: str, install_dir: str):
 
 
 def build_sora(
-    platform: str, sora_dir: str, sora_args: List[str], debug: bool, webrtc_build_dir: Optional[str]
+    platform: str,
+    local_sora_cpp_sdk_dir: str,
+    local_sora_cpp_sdk_args: List[str],
+    debug: bool,
+    local_webrtc_build_dir: Optional[str],
 ):
-    if debug and "--debug" not in sora_args:
-        sora_args = ["--debug", *sora_args]
-    if webrtc_build_dir is not None:
-        sora_args = ["--webrtc-build-dir", webrtc_build_dir, *sora_args]
+    if debug and "--debug" not in local_sora_cpp_sdk_args:
+        local_sora_cpp_sdk_args = ["--debug", *local_sora_cpp_sdk_args]
+    if local_webrtc_build_dir is not None:
+        local_sora_cpp_sdk_args = [
+            "--local-webrtc-build-dir",
+            local_webrtc_build_dir,
+            *local_sora_cpp_sdk_args,
+        ]
 
-    with cd(sora_dir):
-        cmd(["python3", "run.py", platform, *sora_args])
+    with cd(local_sora_cpp_sdk_dir):
+        cmd(["python3", "run.py", platform, *local_sora_cpp_sdk_args])
 
 
 class SoraInfo(NamedTuple):
@@ -766,11 +890,11 @@ class SoraInfo(NamedTuple):
 
 
 def get_sora_info(
-    platform: str, sora_dir: Optional[str], install_dir: str, debug: bool
+    platform: str, local_sora_cpp_sdk_dir: Optional[str], install_dir: str, debug: bool
 ) -> SoraInfo:
-    if sora_dir is not None:
+    if local_sora_cpp_sdk_dir is not None:
         configuration = "debug" if debug else "release"
-        install_dir = os.path.join(sora_dir, "_install", platform, configuration)
+        install_dir = os.path.join(local_sora_cpp_sdk_dir, "_install", platform, configuration)
 
     return SoraInfo(
         sora_install_dir=os.path.join(install_dir, "sora"),
@@ -810,6 +934,16 @@ def install_rootfs(version, install_dir, conf):
     link = os.path.join(rootfs_dir, "usr", "lib", "aarch64-linux-gnu", "tegra", "libnvbuf_fdmap.so")
     file = os.path.join(
         rootfs_dir, "usr", "lib", "aarch64-linux-gnu", "tegra", "libnvbuf_fdmap.so.1.0.0"
+    )
+    if os.path.exists(file) and not os.path.exists(link):
+        os.symlink(os.path.basename(file), link)
+
+    # JetPack 6 から tegra → nvidia になった
+    link = os.path.join(
+        rootfs_dir, "usr", "lib", "aarch64-linux-gnu", "nvidia", "libnvbuf_fdmap.so"
+    )
+    file = os.path.join(
+        rootfs_dir, "usr", "lib", "aarch64-linux-gnu", "nvidia", "libnvbuf_fdmap.so.1.0.0"
     )
     if os.path.exists(file) and not os.path.exists(link):
         os.symlink(os.path.basename(file), link)
@@ -930,8 +1064,6 @@ def install_sdl2(
         ]
         if platform == "windows":
             cmake_args += [
-                "-G",
-                "Visual Studio 16 2019",
                 "-DSDL_FORCE_STATIC_VCRT=ON",
                 "-DHAVE_LIBC=ON",
             ]
@@ -1497,32 +1629,32 @@ def get_webrtc_platform(platform: Platform) -> str:
 # `--sora-args '--test'` のようにスペースを使うと、ハイフンから始まるオプションが正しく解釈されない
 def add_sora_arguments(parser):
     parser.add_argument(
-        "--sora-dir",
+        "--local-sora-cpp-sdk-dir",
         type=os.path.abspath,
         default=None,
         help="Refer to local Sora C++ SDK. "
         "When this option is specified, Sora C++ SDK will also be built.",
     )
     parser.add_argument(
-        "--sora-args",
+        "--local-sora-cpp-sdk-args",
         type=shlex.split,
         default=[],
-        help="Options for building local Sora C++ SDK when `--sora-dir` is specified.",
+        help="Options for building local Sora C++ SDK when `--local-sora-cpp-sdk-dir` is specified.",
     )
 
 
 # add_sora_arguments と同様の注意点があるので注意すること
 def add_webrtc_build_arguments(parser):
     parser.add_argument(
-        "--webrtc-build-dir",
+        "--local-webrtc-build-dir",
         type=os.path.abspath,
         default=None,
         help="Refer to local webrtc-build. "
         "When this option is specified, webrtc-build will also be built.",
     )
     parser.add_argument(
-        "--webrtc-build-args",
+        "--local-webrtc-build-args",
         type=shlex.split,
         default=[],
-        help="Options for building local webrtc-build when `--webrtc-build-dir` is specified.",
+        help="Options for building local webrtc-build when `--local-webrtc-build-dir` is specified.",
     )