+++ examples/messaging_recvonly_sample/ubuntu-22.04_x86_64/run.py
@@ -36,16 +36,16 @@ def install_deps(
     build_dir,
     install_dir,
     debug,
-    webrtc_build_dir: Optional[str],
-    webrtc_build_args: List[str],
-    sora_dir: Optional[str],
-    sora_args: List[str],
+    local_webrtc_build_dir: Optional[str],
+    local_webrtc_build_args: List[str],
+    local_sora_cpp_sdk_dir: Optional[str],
+    local_sora_cpp_sdk_args: List[str],
 ):
     with cd(BASE_DIR):
         version = read_version_file("VERSION")
 
         # WebRTC
-        if webrtc_build_dir is None:
+        if local_webrtc_build_dir is None:
             install_webrtc_args = {
                 "version": version["WEBRTC_BUILD_VERSION"],
                 "version_file": os.path.join(install_dir, "webrtc.version"),
@@ -57,15 +57,17 @@ def install_deps(
         else:
             build_webrtc_args = {
                 "platform": "ubuntu-22.04_x86_64",
-                "webrtc_build_dir": webrtc_build_dir,
-                "webrtc_build_args": webrtc_build_args,
+                "local_webrtc_build_dir": local_webrtc_build_dir,
+                "local_webrtc_build_args": local_webrtc_build_args,
                 "debug": debug,
             }
             build_webrtc(**build_webrtc_args)
 
-        webrtc_info = get_webrtc_info("ubuntu-22.04_x86_64", webrtc_build_dir, install_dir, debug)
+        webrtc_info = get_webrtc_info(
+            "ubuntu-22.04_x86_64", local_webrtc_build_dir, install_dir, debug
+        )
 
-        if webrtc_build_dir is None:
+        if local_webrtc_build_dir is None:
             webrtc_version = read_version_file(webrtc_info.version_file)
 
             # LLVM
@@ -91,10 +93,16 @@ def install_deps(
             install_llvm(**install_llvm_args)
 
         # Sora C++ SDK, Boost
-        if sora_dir is None:
+        if local_sora_cpp_sdk_dir is None:
             install_sora_and_deps("ubuntu-22.04_x86_64", source_dir, install_dir)
         else:
-            build_sora("ubuntu-22.04_x86_64", sora_dir, sora_args, debug, webrtc_build_dir)
+            build_sora(
+                "ubuntu-22.04_x86_64",
+                local_sora_cpp_sdk_dir,
+                local_sora_cpp_sdk_args,
+                debug,
+                local_webrtc_build_dir,
+            )
 
         # CMake
         install_cmake_args = {
@@ -139,10 +147,10 @@ def main():
         build_dir,
         install_dir,
         args.debug,
-        args.webrtc_build_dir,
-        args.webrtc_build_args,
-        args.sora_dir,
-        args.sora_args,
+        args.local_webrtc_build_dir,
+        args.local_webrtc_build_args,
+        args.local_sora_cpp_sdk_dir,
+        args.local_sora_cpp_sdk_args,
     )
 
     configuration = "Debug" if args.debug else "Release"
@@ -151,9 +159,9 @@ def main():
     mkdir_p(sample_build_dir)
     with cd(sample_build_dir):
         webrtc_info = get_webrtc_info(
-            "ubuntu-22.04_x86_64", args.webrtc_build_dir, install_dir, args.debug
+            "ubuntu-22.04_x86_64", args.local_webrtc_build_dir, install_dir, args.debug
         )
-        sora_info = get_sora_info(platform, args.sora_dir, install_dir, args.debug)
+        sora_info = get_sora_info(platform, args.local_sora_cpp_sdk_dir, install_dir, args.debug)
 
         cmake_args = []
         cmake_args.append(f"-DCMAKE_BUILD_TYPE={configuration}")