+++ include/sora/sora_signaling.h
@@ -15,8 +15,9 @@
 #include <api/peer_connection_interface.h>
 #include <api/scoped_refptr.h>
 
-#include "data_channel.h"
-#include "websocket.h"
+#include "sora/data_channel.h"
+#include "sora/version.h"
+#include "sora/websocket.h"
 
 namespace sora {
 
@@ -79,6 +80,7 @@ struct SoraSignalingConfig {
   boost::json::value video_vp9_params;
   boost::json::value video_av1_params;
   boost::json::value video_h264_params;
+  boost::json::value video_h265_params;
   std::string audio_streaming_language_code;
   boost::json::value metadata;
   boost::json::value signaling_notify_metadata;
@@ -133,6 +135,8 @@ struct SoraSignalingConfig {
   rtc::PacketSocketFactory* socket_factory = nullptr;
 
   bool disable_signaling_url_randomization = false;
+
+  boost::optional<http_header_value> user_agent;
 };
 
 class SoraSignaling : public std::enable_shared_from_this<SoraSignaling>,